django_caroma/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==25.3.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
